FINISHED!!!! :DDD

+++ controllers/articleController.js
@@ -54,10 +54,19 @@ module.exports.displayAll = async function(req, res) {
 
 module.exports.renderEditForm = async function(req, res) {
     const article = await Article.findByPk(req.params.articleId);
+    if (!article.isOwnedBy(user)) {
+        res.redirect('/');
+        return;
+    }
     res.render('articles/edit', {article});
 };
 
 module.exports.updateArticle = async function(req, res) {
+    const article = await Article.findByPk(req.params.articleId);
+    if (!article.isOwnedBy(user)) {
+        res.redirect('/');
+        return;
+    }
     await Article.update({
         title: req.body.title,
         intro: req.body.intro,
@@ -72,6 +81,11 @@ module.exports.updateArticle = async function(req, res) {
 };
 
 module.exports.deleteArticle = async function(req, res) {
+    const article = await Article.findByPk(req.params.articleId);
+    if (!user.is('admin') && !article.isOwnedBy(user)) {
+        res.redirect('/');
+        return;
+    }
     await Article.destroy({
         where: {
             id: req.params.articleId
